backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
polygon-api-client==1.12.3
supabase==2.3.1
